+++ .gitignore
@@ -15,7 +15,7 @@
 /src/views/somns.html
 /src/views/test.html
 /tests/data/actual-results
-/tests/large-payload.json
+/tests/data/large-payload.json
 /tmp
 test.figures
 test.html
