+++ src/backend/rebench/results.ts
@@ -6,7 +6,7 @@ import { Database } from '../db/db.js';
 import { createValidator } from './api-validator.js';
 import { DEBUG } from '../util.js';
 import { log } from '../logging.js';
-import { completeRequest, startRequest } from '../../perf-tracker.js';
+import { completeRequest, startRequest } from '../perf-tracker.js';
 
 const validateFn: ValidateFunction = DEBUG ? createValidator() : <any>undefined;
