+++ tests/backend/compare/db-data.test.ts
@@ -1,16 +1,16 @@
 import { describe, expect, it } from '@jest/globals';
 import { readFileSync } from 'node:fs';
-import { robustPath } from '../src/backend/util.js';
-import { collateMeasurements } from '../src/backend/compare/db-data.js';
+import { robustPath } from '../../../src/backend/util.js';
+import { collateMeasurements } from '../../../src/backend/compare/db-data.js';
 import {
   ResultsByBenchmark,
   ResultsByExeSuiteBenchmark,
   ResultsBySuiteBenchmark
-} from '../src/backend/compare/prep-data.js';
+} from '../../../src/backend/compare/prep-data.js';
 import type {
   MeasurementData,
   ProcessedResult
-} from '../src/backend/db/types.js';
+} from '../../../src/backend/db/types.js';
 
 const dataJsSOM = JSON.parse(
   readFileSync(
