+++ tests/shared/data-format.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@jest/globals';
-import type { Environment } from '../src/backend/db/types.js';
+import type { Environment } from '../../src/backend/db/types.js';
 import {
   asHumanHz,
   asHumanMem,
@@ -9,8 +9,8 @@ import {
   per,
   r0,
   r2
-} from '../src/shared/data-format.js';
-import { DataSeriesVersionComparison } from '../src/views/view-types.js';
+} from '../../src/shared/data-format.js';
+import { DataSeriesVersionComparison } from '../../src/views/view-types.js';
 
 describe('Format Functions for Numerical Values', () => {
   describe('r0 - round to 0 decimal places', () => {
